Added jwt auth

+++ backend/controllers/user-controller.js
@@ -4,22 +4,37 @@ import { setResponse, setErrorResponse } from './response-handler.js';
 // Method to get all users
 export const get = async (request, response) => {
     try {
+        const accessToken = request.headers.authorization.split(' ')[1];
+        userService.verifyAccessToken(accessToken);
         const users = await userService.getAll();
         console.log(users);
         setResponse(users, response, 200);
-    } catch (err) {
-        console.log(err);
-        setErrorResponse(err, response);
+    } 
+    catch (err) 
+    {
+        if(err.message === 'JWT token expired')
+        {
+            response.status(401).json({ error: 'JWT token expired' });
+        }
+        else
+        {
+            console.log(err);
+            setErrorResponse(err, response);
+        }  
     }
 }
 
 // Method to post or save the user
 export const post = async (request, response) => {
     try {
+        // const accessToken = request.headers.authorization.split(' ')[1];
+        // userService.verifyAccessToken(accessToken);
         const newUser = { ...request.body };
         const user = await userService.save(newUser);
         setResponse(user, response, 200);
-    } catch (err) {
+    } 
+    catch (err) 
+    {
         console.log(err);
         setErrorResponse(err, response);
     }
@@ -28,10 +43,14 @@ export const post = async (request, response) => {
 // Method to get an user by id
 export const getById = async (request, response) => {
     try {
+        const accessToken = request.headers.authorization.split(' ')[1];
+        userService.verifyAccessToken(accessToken);
         const id = request.params.id;
         const user = await userService.getById(id);
         setResponse(user, response, 200);
-    } catch (err) {
+    } 
+    catch (err) 
+    {
         console.log(err);
         setErrorResponse(err, response);
     }
@@ -40,39 +59,90 @@ export const getById = async (request, response) => {
 // Method to get users by params
 export const getByParams = async (request, response) => {
     try {
+        const accessToken = request.headers.authorization.split(' ')[1];
+        userService.verifyAccessToken(accessToken);
         const params = { ...request.query };
         const users = await userService.getByParams(params);
         setResponse(users, response, 200);
-    } catch (err) {
-        console.log(err);
-        setErrorResponse(err, response);
+    } 
+    catch (err) 
+    {
+        if(err.message === 'JWT token expired')
+        {
+            response.status(401).json({ error: 'JWT token expired' });
+        }
+        else
+        {
+            console.log(err);
+            setErrorResponse(err, response);
+        }  
     }
 }
 
 // Method to update the user
 export const update = async (request, response) => {
     try {
+        const accessToken = request.headers.authorization.split(' ')[1];
+        const requestingUser = userService.verifyAccessToken(accessToken);
+        console.log(requestingUser);
+        console.log(request.params.id);
+        if(requestingUser.id !== request.params.id) 
+        {
+            throw new Error('You are not authorized to update this user');
+        }
         // Getting id from request
         const id = request.params.id;
         const updatedUser = { ...request.body };
         const user = await userService.update(updatedUser, id);
         setResponse(user, response, 200);
-    } catch (err) {
-        console.log(err);
-        setErrorResponse(err, response)
+    } 
+    catch (err) 
+    {
+        if (err.message === 'You are not authorized to update this user') {
+            response.status(401).json({ error: 'You are not authorized to update this user' });
+        }
+        else if(err.message === 'JWT token expired')
+        {
+            response.status(401).json({ error: 'JWT token expired' });
+        }
+        else
+        {
+            console.log(err);
+            setErrorResponse(err, response);
+        } 
     }
 }
 
 // Method to delete an user
 export const remove = async (request, response) => {
     try {
+        const accessToken = request.headers.authorization.split(' ')[1];
+        const requestingUser = userService.verifyAccessToken(accessToken);
+        console.log(requestingUser);
+        console.log(request.params.id);
+        if(requestingUser.id !== request.params.id) 
+        {
+            throw new Error('You are not authorized to delete this user');
+        }
         // Getting id from request
-        const id = request.params.id;
+        const id = request.params.id; 
         const removedUser = await userService.remove(id);
         setResponse(removedUser, response, 200);
-    } catch (err) {
-        console.log(err);
-        setErrorResponse(err, response)
+    } 
+    catch (err) 
+    {
+        if (err.message === 'You are not authorized to delete this user') {
+            response.status(401).json({ error: 'You are not authorized to delete this user' });
+        }
+        else if(err.message === 'JWT token expired')
+        {
+            response.status(401).json({ error: 'JWT token expired' });
+        }
+        else
+        {
+            console.log(err);
+            setErrorResponse(err, response);
+        } 
     }
 }
 
@@ -80,13 +150,19 @@ export const remove = async (request, response) => {
 export const loginUser = async (request, response) => {
     try {
         const user = { ...request.body };
-        const validatedUser = await userService.login(user);
-        setResponse(validatedUser, response, 200);
-    } catch (err) {
-        if (err.message === 'User not found') {
+        const { user: validatedUser, accessToken, refreshToken } = await userService.login(user);
+
+        setResponse({ user: validatedUser, accessToken, refreshToken }, response, 200);
+    } 
+    catch (err) 
+    {
+        if (err.message === 'User not found') 
+        {
             // Handle the case where the user is not found
             response.status(404).json({ error: 'User not found' });
-        } else {
+        } 
+        else 
+        {
             // Handle other errors
             console.log(err);
             setErrorResponse(err, response);
@@ -94,3 +170,46 @@ export const loginUser = async (request, response) => {
     }
 }
 
+//Logout user
+export const logout = async (request, response) => {
+    try {
+      const refreshToken = request.body.token;
+      await userService.verifyRefreshToken(refreshToken);
+      
+      const accessToken = request.headers.authorization.split(' ')[1];
+      const user = userService.verifyAccessToken(accessToken);
+      await userService.logout(user.id);
+      setResponse('Logout successful', response, 200);
+    } 
+    catch (err) 
+    {
+      console.log(err);
+      setErrorResponse(err, response);
+    }
+  };
+
+  //Get new access token
+  export const refreshTokens = async (request, response) => {
+    try {
+      const refreshToken = request.body.token;
+      const tokens = await userService.refreshTokens(refreshToken);
+  
+      setResponse(tokens, response, 200);
+    } 
+    catch (err) 
+    {
+        if(err === 'Refresh token is required')
+        {
+            response.status(401).json({error: 'Refresh token is required'});
+        }
+        else if(err === 'Refresh token is not valid')
+        {
+            response.status(403).json({error: 'Refresh token is not valid'});
+        }
+        else
+        {
+            console.log(err);
+            setErrorResponse(err, response);
+        }
+    }
+  };
